pyproject.toml
[project]
name = "memes-ranker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "aiosqlite>=0.21.0",
    "coolname>=2.2.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qrcode[pil]>=8.2",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest-cov>=6.2.1",
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["test_basic_flow.py"]
